requirements.txt
asgiref==3.7.2
Django==4.2.4
django-model-utils==4.3.1
djangorestframework==3.14.0
MuPDF
PyMuPDF==1.22.5
Pillow==10.0.0
pytz==2023.3
sqlparse==0.4.4
psycopg2